### 2024-09-16 #7230 @editorialbot
[PRE REVIEW]: Mozzie: a computationally efficient simulator for the spatio-temporal modelling of mosquitoes
**Submitting author:** @WilkAndy (<a href="http://orcid.org/0000-0001-6472-9560">Andy Wilkins</a>)
**Repository:** https://github.com/csiro-risk-assessment/mozzie
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
