### 2024-09-16 #7230 @Kevin-Mattheus-Moerman
@mengqi-z I was wondering if you could help edit this submission. It may be somewhat off topic, but I figured your geospatial research expertise may be applicable.
