### 2024-09-16 #7230 @WilkAndy
Confirming that:

```
- 10.1101/2023.09.09.556958 may be a valid DOI for title: MGDrivE 3: A decoupled vector-human framework for ...
- 10.1080/17513750902803588 may be a valid DOI for title: A note on the nonautonomous delay Beverton–Holt mo...
```

are correct DOIs.
