### 2024-09-23 #7230 @mengqi-z
👋 @kerguler @dinacmistry @GilChrist19 @yafeiwang89 - Would any of you be interested and available to review this submission for JOSS? The review process is straightforward, handled through GitHub issues, and follows a checklist-driven approach: https://joss.readthedocs.io/en/latest/review_criteria.html

If you are unable to take on this review, I would greatly appreciate any recommendations for other potential reviewers with relevant expertise.
