### 2024-09-26 #7230 @mengqi-z
👋 @jkbren @emilydolson @chryswoods @bramvandijk88 - Would you be willing and available to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
