### 2024-09-30 #7230 @mengqi-z
👋 @agmunozs @SimonbJohnson @agerada @kathlandgren - Would any of you be available to review this submission for JOSS? Given your background, I believe you'd be a great fit to review this paper.

JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html

If you're unable to review, any recommendations for other potential reviewers with relevant expertise would also be greatly appreciated. Thank you!
