### 2024-10-01 #7230 @agmunozs
Hi, sorry. Too busy at the momentt!
>
>
>
> Ángel .G Muñoz, PhD (he/him)
>
> Senior Scientist | Ramón y Cajal Fellow
> Leader - Climate Services Team (CST)
> Earth System Services <https://ess.bsc.es/> Group | Earth Sciences Department
> Barcelona Supercomputing Center - Centro Nacional de Supercomputación (BSC-CNS)
> Address: C/ Jordi Girona, 31, 08034 Barcelona, Spain | Torre Girona, 2nd Floor
> Personal Website:  Ángel G. Muñoz site <https://www.bsc.es/munoz-solorzano-angel-garikoitz>Email:  ***@***.***
> Twitter:  @agmunozs <https://twitter.com/agmunozs?lang=es>

> On 30 Sep 2024, at 19:54, Mengqi Zhao ***@***.***> wrote:
>
>
> 👋 @agmunozs <https://github.com/agmunozs> @SimonbJohnson <https://github.com/SimonbJohnson> @agerada <https://github.com/agerada> @kathlandgren <https://github.com/kathlandgren> - Would any of you be available to review this submission for JOSS? Given your background, I believe you'd be a great fit to review this paper.
>
> JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> If you're unable to review, any recommendations for other potential reviewers with relevant expertise would also be greatly appreciated. Thank you!
>
