### 2024-10-02 #7230 @mengqi-z
👋 @VincentStimper @jabirali @zoometh - Would you be interested in reviewing this submission for JOSS?

JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
