### 2024-10-07 #7324 @mengqi-z
👋🏼 @emilydolson, @slwu89 - This is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```

Please feel free to ping me (`@mengqi-z`) if you have any questions/concerns.
