### 2024-10-14 csiro-risk-assessment/mozzie#48 @slwu89
JOSS review comments
Linking to JOSS review thread: https://github.com/openjournals/joss-reviews/issues/7324

- [x] Regarding "General Checks: Reproducibility": is Figure 1 in the paper from code that exists in this (mozzie) repo? Or a different repo? Either way it would be nice to have a link to code to reproduce, if it exists.
- [x] Regarding "Documentation: A statement of need": the README, which seems to be the main point of documentation for the software, is lacking a high level description of what the software does and whom might be interested in it.
- [x] Regarding "Documentation: Functionality documentation": while the README does contain information about the main classes and methods there is not sufficient information about how to set up the simulation loop (i.e. order of updating, time step, etc), and there is not information regarding how to use the auxiliary (plotting) code. This could be added to the README, but as it is quite long perhaps a dedicated docs website would be more appropriate for a project of this complexity and scale.
- [x] Regarding "Documentation: Community guidelines": there is insufficient information on points 2) Report issues or problems with the software 3) Seek support
- [x] Regarding "Software Paper: Statement of need" and "State of field": I think that the paper needs to better separate the sections Summary, Statement of need, and State of field. Currently the State of the field section is contained in the Statement of need section. Furthermore the Statement of need section could be more clear about who the anticipated users of the software are.
