### 2024-10-14 d4bcb0d
Add description to README.md : point 2 of @slwu89 .   Refs #48
