### 2024-10-14 ba7b4af
Add info about setting up the simulation : point 3 of @slwu89 .  Refs #48
