+++ README.md
@@ -139,24 +139,145 @@ Before attempting a full-scale simulation with complicated lifecycle dynamics, y
 Each of these python files contains extensive in-code documentation.  Run each of them using, for instance, `python3 logistic.py`.
 
 
-## Simulating in general
+## How to set up a simulation
 
-The core code consists of python objects that you must instantiate in a "runner" python script that defines your mathematical model.  An example is `example1/runner.py`, and there are other more sophisticated models in the other `example` directories.  Run the simulations using, for example,
+The core code consists of python objects that you must instantiate in a "runner" python script that defines your mathematical model.  The aforementioned primers contain simple examples, and now we describe `example1/runner.py` that contains all the components of a full-scale simulation.  There are other more sophisticated models in the other `example*` directories.  Run the simulations using, for example,
 
 `python3 runner.py`
 
-Generally, each "runner" python script contains the following.
+### runner.py: the `import` block
 
-- An `import` block in which you import all the core libraries (described below) and any other python libraries you need.
-- A block in which you set up the grid and active cells, defining the spatial extents and discretisation of the model.  You will use `Grid` and `Grid.setActiveAndInactive`.
-- A block that defines `Wind` (multiple input files corresponding to different times).  This is a little different from the generic spatially-varying parameters (next item) because it usually requires some sort of pre-processing, or is read from specially-preprocessed files.
-- A block in which you read files corresponding to spatially-varying parameters in your model, such as carrying capacities.  You'll typically have to read multiple of these per parameter, as the parameters will also vary with time.  You will use `SpatialDependence` and `SpatialDependence.restrictToActive`
-- A definition of your cell dynamics (the ODE model), such as `CellDynamicsMosquito23`
-- A block that defines the populations and parameters, and sets initial conditions.  You'll use `PopulationsAndParameters`.
-- A block that gathers the grid structure, the populations and the cell dynamics together into a `SpatialDynamics` object.
-- A block that controls the time-stepping of the simulation, involving `diffuse`, `evolveCells`, and `advect`.  There can be multiple of these, depending on the complexity of your model, as different carrying capacities, wind vectors, etc, are used at different times.
+The `import` block is used to import all the core libraries (described below) and any other python libraries you need.  `example1/runner.py` looks like:
 
-Because of this "block" structure, you can run partial simulations, for instance, just processing wind files, or loading files, processing in some way and outputting to produce figures.  You can also chain together multiple simulations that rely on just one initial block of file reading, to avoid reading data files for each and every simulation.
+```
+import os
+import sys
+import array
+import timeit
+
+# so we can find our ../code no matter how we are called
+findbin = os.path.dirname(os.path.realpath(sys.argv[0]))
+sys.path.append(findbin + "/../code")
+
+from wind import Wind
+from grid import Grid
+from cellDynamics import CellDynamicsLogistic1_1
+from spatialDynamics import SpatialDynamics
+from spatialDependence import SpatialDependence
+from populationsAndParameters import PopulationsAndParameters
+```
+
[…]
