### 2024-10-14 baeacdd
Add info about seeking help and reporting issues : point 4 of @slwu89 .  Refs #48
