### 2024-10-14 4d2b458
In paper.md: separate statement of need from state of the field, and add likely user base : point 5 of @slwu89 .  Refs #48
