+++ paper.md
@@ -63,9 +63,7 @@ AN ALTERATE PARAGRAPH:
 `Mozzie` enables simulation of the lifecycle and spatial spread of mosquitoes.  Mosquitoes transmit malaria, and `Mozzie` is designed to support risk assessments associated with genetic alterations of mosquitoes aimed at eliminating malaria.  Treatment of the mosquito lifecycle includes competition and breeding between multiple (sub-)species, allowing researchers to assess potential transfer of the genetic alteration to other (sub-)species.  To assess the spatial spread of wild-type and genetically-altered mosquitoes, `Mozzie` can simulate the local diffusion of individuals through the landscape, and long-distance dispersal via wind.  `Mozzie` can be used for risk assessments at the local, regional or continental scale.  For example, wind patterns and the capacity of the landscape to support mosquitoes can vary spatially and temporally, reflecting seasonality, local conditions and/or daily variations, etc.  Genetic alterations can be flexibly modelled with any number of alleles and the capacity for Mendelian or non-Mendelian inheritance, including gene drives.
 -->
 
-# Statement of need
-
-The numerical implementation of Mozzie is ecologically interpretable [@Hosack2023] and computationally and I/O efficient.  This allows rapid simulation at continental scales to investigate sensitivity to input parameters, as required in risk assessments.  It is written in [Cython](https://cython.org) [@behnel2010cython] (a mixture of Python and C), and simulations are run using Python. The test coverage of the `Mozzie` codebase is over 99%, meaning it is also suitable for risk assessments that could be subject to considerable scrutiny.
+# State of the field
 
 Alternatives to `Mozzie` include:
 
@@ -77,9 +75,15 @@ Alternatives to `Mozzie` include:
 - The [exDE](https://dd-harp.github.io/exDE/) R package solves models of mosquito-borne pathogen dynamics and control [@Wu2023].  Attention is payed to sophisticated representations of mosquito lifecycles, including exogenous forcing by weather and vector control, as well as mosquito-malaria-human interatctions.  Although a single mosquito species is the focus, the framework allows for multiple species.  The code centers on traditional vector controls, rather than genetic controls.
 - Berkeley's [MGDrivE](https://marshalllab.github.io/MGDrivE/) is an open-source framework to study gene-drives in mosquito populations [@mgdrive2019;@mondal2024mgdrive], which is written in R.  With regards to lifecycle dynamics, MGDrivE has similar functionality to `Mozzie`, although MGDrivE focuses on single species, in contrast to `Mozzie` where transfer of genetic constructs between (sub-)species is of interest.  MGDrivE's spatial structure is based on a network, where each node in the network could be thought of as a household, house block, or even a city.  In terms of functionality, MGDrivE is the most similar to `Mozzie`, although the numerical methods employed are quite different.
 
-If spatially explicit, these model spatial structure using a network.  In contrast, `Mozzie` uses a continuous-space (diffusion-advection equation) approach, deliberately incorporating long-range dispersal in a way that is ecologically interpretable [@Hosack2023].  In addition, `Mozzie` does not focus on single species, but concentrates on the interaction of multiple (sub-)species.  Many of the aforementioned alternatives contain human agents and the malaria parasite, which `Mozzie` does not.
[…]
