### 2024-10-15 60691cb
In paper.md: add link to DAP to produce Fig1 : point 1 of @slwu89 .  Refs #48
