### 2024-10-15 f06c8b9
In paper.md: clearer statement of need : point 5 of @slwu89 .  Refs #48
