### 2024-10-15 csiro-risk-assessment/mozzie#48 @WilkAndy
Hi @slwu89 - i believe i've addressed all your concerns.  The updated README.md is in the repo, and paper.pdf can be found here:  https://github.com/csiro-risk-assessment/mozzie/actions/runs/11336937388/artifacts/2055601183 .

Could you please confirm that the changes are suitable, or give me more instructions on how to make things better?
