### 2024-10-16 csiro-risk-assessment/mozzie#48 @slwu89
Hi @WilkAndy yes it looks like all my comments were suitably addressed. One extremely minor thing, on line 14 of the (new) paper, "genetic" has a type ("genetically" was probably intended).
