+++ paper.md
@@ -48,7 +48,7 @@ See [instructions](https://joss.readthedocs.io/en/latest/paper.html) for instruc
 
 `Mozzie` enables simulation of the lifecycle and spatial spread of mosquitoes.  `Mozzie` can be used to assess risks associated with disease-control strategies at local, regional or continental scales.  Most particularly, strategies involving genetic alterations of mosquitoes to eliminate malaria, are of prime interest.
 
-More technically, `Mozzie` simulates a population-dynamics model that uses differential equations or delay differential equations [@BOHNER2018114;@El-Hachem2024] to describe the spread and persistence of mosquitoes that may be genetic altered.  Genetic alterations are flexibly modelled: these can involve any number of alleles; Mendelian or non-Mendelian inheritance, including gene drives; they can be self-limiting or self-sustaining; and can include the emergence of resistant allelles.  The model allows simulation of $N$ mosquito species.   It incorporates mate-choice, hybridisation and intra-specific competition that occur within complexes of mosquito species [@BEETON2020110072].  This fills a gap that currently exists among similar models, allowing researchers to assess potential transfer of the genetic alterations between (sub-)species.
+More technically, `Mozzie` simulates a population-dynamics model that uses differential equations or delay differential equations [@BOHNER2018114;@El-Hachem2024] to describe the spread and persistence of mosquitoes that may be genetically altered.  Genetic alterations are flexibly modelled: these can involve any number of alleles; Mendelian or non-Mendelian inheritance, including gene drives; they can be self-limiting or self-sustaining; and can include the emergence of resistant allelles.  The model allows simulation of $N$ mosquito species.   It incorporates mate-choice, hybridisation and intra-specific competition that occur within complexes of mosquito species [@BEETON2020110072].  This fills a gap that currently exists among similar models, allowing researchers to assess potential transfer of the genetic alterations between (sub-)species.
 
 `Mozzie` supports spatial and temporal variations in lifecyle parameters, and local diffusion and wind-assisted, long range, advection.  For example, wind patterns and the capacity of the landscape to support mosquitoes can vary spatially and temporally, reflecting daily variations, seasonality, and local conditions.
 
