### 2024-10-16 #7324 @WilkAndy
The current version of the paper is here: https://github.com/csiro-risk-assessment/mozzie/actions/runs/11357812298/artifacts/2061254076
