### 2024-10-16 297b48e
remove CC BY sentence from acknowledgements of paper.md.  Fixes #49
