### 2024-11-26 #7324 @bramvandijk88
> 👋 @bramvandijk88 - I hope you're doing well. Would you be available to review a submission for JOSS? I initially reached out a while back, and you mentioned you might have availability later this fall. Unfortunately, one of our reviewers is no longer able to review, and I was hoping you could step in to fill the spot. Let me know if this works for you. Thanks!

Hi! Sorry, I’ve overcommitted task-wise, so I’m going to have to say no. I’ll keep trying to review 1-2 submissions for JOSS every year :)
