### 2024-11-26 #7324 @mengqi-z
👋 @martibosch @chrisvoncsefalvay - Would any of you be interested in and able to take on reviewing this JOSS submission? JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
