### 2024-12-03 #7324 @mengqi-z
👋 @jGaboardi @jedalong - Would any of you be available to review this submission for JOSS? We carries out a checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
