### 2024-12-03 #7324 @mengqi-z
👋 @klmedeiros - Would you be interested in reviewing this submission for JOSS? We follow a check-list driven approach: https://joss.readthedocs.io/en/latest/review_criteria.html
