### 2024-12-05 #7324 @klmedeiros
@mengqi-z thank you for offering, i appreciate it! unfortunately i don't know C well enough to review this the way it deserves. thanks again!
