### 2024-12-05 #7324 @mengqi-z
👋 @stulacy - Would you be willing to review this JOSS submission? The review process follows a check-list driven approach: https://joss.readthedocs.io/en/latest/review_criteria.html
