### 2024-12-17 #7324 @mengqi-z
👋 @MatthewKhouzam - Would you be interested in reviewing this submission for JOSS? JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
