### 2025-01-28 csiro-risk-assessment/mozzie#50 @stulacy
Automatically install dependencies
It would help smooth the installation process, as well as making the package more robust if dependencies were explicitly stated in code and installed. At minimum they can be hardcoded into `setup.py`, but having them in a `requirements.txt` which is read by `setup.py` would be more idiomatic and provide a single source of truth for all dependencies and their versions.

Ref https://github.com/openjournals/joss-reviews/issues/7324
