### 2025-01-28 csiro-risk-assessment/mozzie#51 @stulacy
Make Figure 1 in the JOSS paper reproducible
Figure 1 in the JOSS paper displays example simulation outputs. It links to a [repository](https://doi.org/10.25919/rgqc-7520) containing the code and data to reproduce these results, however, the code seems to be setup for specific users containing hardcoded paths to a home directory (`/home/bee069/mozzie/code`), as well as code specific to a Slurm HPC.

Is it possible to provide example code that can be run by anyone and doesn't need this specific HPC setup?

Ref https://github.com/openjournals/joss-reviews/issues/7324
