### 2025-01-28 csiro-risk-assessment/mozzie#50 @WilkAndy
Yes, thank you @stulacy .   I will push a commit that includes `requirements.txt`
