### 2025-01-28 3ecb218
add requirements.txt (and modified README.md appropriately).  Fixes #50
