+++ requirements.txt
@@ -1,0 +1,6 @@
+numpy~=2.2
+scipy~=1.15
+Cython~=3.0
+coverage~=7.6
+setuptools~=65.5
+matplotlib~=3.10
