### 2025-01-29 546c8cb
requirements.txt now requires setuptools 70.0 version, to address a security problem in earlier setuptools versions.  Refs #50
