+++ requirements.txt
@@ -2,5 +2,5 @@ numpy~=2.2
 scipy~=1.15
 Cython~=3.0
 coverage~=7.6
-setuptools~=65.5
+setuptools~=70.0
 matplotlib~=3.10
