### 2025-02-20 #7324 @WilkAndy
> 👋 [@stulacy](https://github.com/stulacy) [@WilkAndy](https://github.com/WilkAndy) - It looks like you are making great progresses! Could you both provide a brief update in this thread on how things are going?

Hi @mengqi-z , thanks for the prompt.   We have addressed one of @stulacy 's suggestions, and next week @nickbeeton will attempt to make Figure 1 reproducible.  Note that this latter one is working in a repo that is associated with an already-published paper, which is a different repository to `mozzie`.
