### 2025-02-24 #7324 @stulacy
Hi @mengqi-z , I'm looking forward to being able to reproduce Fig 1 from the paper.
The other thing on my list is to look at the examples in the `primers` directory to verify the functional claims, which I hope to get around to shortly.
