+++ README.md
@@ -6,7 +6,7 @@
 
 More technically, `Mozzie` simulates a population-dynamics model that uses differential equations or delay differential equations to describe the spread and persistence of mosquitoes that may be genetic altered.  Genetic alterations are flexibly modelled: these can involve any number of alleles; Mendelian or non-Mendelian inheritance, including gene drives; they can be self-limiting or self-sustaining; and can include the emergence of resistant allelles.  The model allows simulation of any number of mosquito species.   It incorporates mate-choice, hybridisation and intra-specific competition that occur within complexes of mosquito species.  This fills a gap that currently exists among similar models, allowing researchers to assess potential transfer of the genetic alterations between (sub-)species.
 
-`Mozzie` supports spatial and temporal variations in lifecyle parameters, and local diffusion and wind-assisted, long range, advection.  For example, wind patterns and the capacity of the landscape to support mosquitoes can vary spatially and temporally, reflecting daily variations, seasonality, and local conditions.
+`Mozzie` supports spatial and temporal variations in lifecycle parameters, and local diffusion and wind-assisted, long range, advection.  For example, wind patterns and the capacity of the landscape to support mosquitoes can vary spatially and temporally, reflecting daily variations, seasonality, and local conditions.
 
 Conversely, `Mozzie` does not contain human agents, nor does it consider the effect of genetic control strategies on the prevalence of pathogens such as the malaria parasite, among human or animal populations.
 
@@ -15,16 +15,16 @@ Conversely, `Mozzie` does not contain human agents, nor does it consider the eff
 
 Australia's Commonwealth Scientific and Industrial Research Organisation (CSIRO) welcomes contributions to this code.  The recommended method is to:
  
-1. Submit an Issue on github, describing the reasons for your potential contribution;
-2. Engage in discussions with CSIRO staff on github regarding the best architecture, etc, for your contribution;
-3. When your contribution is ready, issue a Pull Request on github.
+1. Submit an [Issue](https://github.com/csiro-risk-assessment/mozzie/issues) on GitHub, describing the reasons for your potential contribution;
+2. Engage in discussions with CSIRO staff on GitHub regarding the best architecture, etc, for your contribution;
+3. When your contribution is ready, issue a Pull Request on GitHub.
  
 CSIRO does not guarantee to inspect such contributions, nor accept them into the code.
 
 
 ## Licensing and disclaimers
 
-This code is released under a GPLv3 license. To discuss other licenses, please contact CSIRO via github.
+This code is released under a GPLv3 license. To discuss other licenses, please contact CSIRO via GitHub.
  
 Copyright (c) 2024 Commonwealth Scientific and Industrial Research Organisation (CSIRO) ABN 41 687 119 230.
  
@@ -32,9 +32,9 @@ CSIRO does not make any guarantees of accuracy or relevance of the code.  CSIRO 
 
 ## Reporting problems and seeking support
 
-Please report problems using a new github issue.  Please tag someone (eg @WilkAndy) in your issue.  CSIRO will endeavour to fix any problems, but does not guarantee to do so.
+Please report problems using a new GitHub [issue](https://github.com/csiro-risk-assessment/mozzie/issues).  Please tag someone (eg @WilkAndy) in your issue.  CSIRO will endeavour to fix any problems, but does not guarantee to do so.
 
-To seek help, please also use a new github issue, and please tag someone (eg @WilkAndy).  Using a github issue means that future users may benefit from the solution to your problem.  CSIRO will endeavour to help, but does not guarantee to do so.
[…]
