### 2025-02-25 #7324 @stulacy
I've run the examples and they all work as described.
I've made a PR and some issues in the repo with suggestions for improving the documentation.
