### 2025-03-02 8336276
Comments on where the code is built, and manipulating PATH, and cd to primers or example directory.  Fixes #53
