### 2025-03-02 4316705
Removed reference to Pearcey in README.md.  Fixes #54
