### 2025-03-17 csiro-risk-assessment/mozzie#51 @nickbeeton
Sorry @stulacy and @WilkAndy for the long delay on this! I've been working steadily trying to reproduce Figure 1, which has been slow due to losing some files and having to rerun large chunks of the process. On the plus side I've been able to double check that the code works! I've got the code running now, and after many false starts I'm fairly confident it should work this time, so all going well this should be done this week.

Note that running the model requires a lot of computation and a lot of memory and I/O use so it may not be feasible on a non-HPC system. I'll make sure there are non-HPC equivalents of all code and remove the hardcoded paths though. I'll also produce intermediate files so that anyone can at least plot Figure 1 given some small output files from the model run.
