### 2025-04-30 #7324 @WilkAndy
Hello @mengqi-z and @stulacy .   Thank you for your patience over the past few weeks as we’ve been revisiting the reproducibility of Figure 1.  This figure was taken from an earlier model simulation, that was published in another paper, and, unfortunately, the figure:
- involves massive data preprocessing prior to running the `mozzie` code;
- is based on a previous, simpler, version of `mozzie`; and
- requires high performance computing to produce.

Because of these three items, we've come to the conclusion that it is a suboptimal demonstration of `mozzie`’s potential.

Our plan is to instead commit a standalone example to our repo.  This example will demonstrate the `mozzie` software platform for users. The example will:
- be self-contained within the `mozzie` repo;
- demonstrate programatically and pictorially the functionality of `mozzie`
- be small enough to run on a personal laptop without access to high performance computing facilities
- require no significant data preprocessing.

We believe this revised approach would improve the paper while directly addressing the @stulacy 's comments above.

The new implementation is requiring us to create this demonstration example and its accompanying figure. We believe that we can incorporate the new approach into an updated `mozzie` repo by **3 June**.    We hope this is satisfactory, @stulacy
