### 2025-06-11 #7324 @WilkAndy
Thank you @mengqi-z .    I believe that I've done all that you asked, above

- The version is v1.0
- The DOI is https://doi.org/10.5281/zenodo.15637648
