### 2025-06-11 #7324 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
