### 2025-06-13 #7324 @WilkAndy
Hi @mengqi-z - we have noticed a typo in our manuscript.  In one place, we wrote _Anopheles coluzzi_ but should have written _Anopheles coluzzii_ .   (Two "i"s at the end.).  Is it possible to change that typo now that the paper is published?

A
