### 2025-06-16 #7324 @mengqi-z
Hi @WilkAndy,

Yes, you can correct the typo in your `paper.md`. Once updated, the track Editor-in-Chief (@Kevin-Mattheus-Moerman) can reaccept the paper. The DOI will remain unchanged.

@Kevin-Mattheus-Moerman - Could you please confirm if this is the correct process and assist @WilkAndy with the update? Thank you!
