### 2025-06-19 #7324 @WilkAndy
Thank you @mengqi-z .  I have fixed the "coluzzii" typo in the paper (via https://github.com/csiro-risk-assessment/mozzie/commit/9e93544801aa9d949e62abf2617a394637a53e02 )  If it is possible to re-accept, that would be wonderful, thank you, @Kevin-Mattheus-Moerman .
